separated the database and main module
