+++ .travis.yml
@@ -1,6 +1,6 @@
 language: python
 python:
-  - '3.7.3'
+  - '3.6.4'
   - "nightly"
 install: 
   - pip install -r requirements.txt
